Reviewed-by: andresin Approved-by: andresin
This squeezes "home" and "advanced" search links into the search bar, replaces the scalable image with a pre-shrunken one that will look good on more browsers.  In order to keep all of this from wrapping on a 600x800 display (the smallest at which everything should basically still look nice), I dropped the form elements down to 0.8em, which is small but not tiny.