Bugs fixed: 90, 91 Reviewed-by: PeterCroome, jgarcke Approved-by: jgarcke
Fix listing tables so that their top margins are respected (by floating them), but remove those top margins on the publisher and imprint pages as they result in doubled margins if kept or insufficient margins in the case of the publisher having no series/imprints if removed.  Marvel and DC both have imprints with no series, which shows that particular problem.

Also put a label on the publisher's URL, which was missing (and looked weird under the two labeled fields, even though its purpose is obvious).  Finally, ensured that the publisher's URL appears both on the series listing and imprint listing pages for the publishers, for consistency with how the other fields are handled.