Bugs: 74, 83 Reviewed-by: jgarcke Approved-by: jgarcke
This isn't as big as it looks!  The model has a bunch of formatting changes, but really just adds the no_* credit flags and the page_count_uncertain flag.

The main templatetag change is to the __credit_visible() method, which now just checks for None and ''.  The no_* credit flags are not currently checked, which is OK for now because the script ensures that the field is always blank if they are set to true.  Those flags will be used in the OI, though.  The rest of the changes here are formatting of comments and removing XML code that I forgot to take out with the other XML code a while ago.

The template changes are for displaying a "?" after the page count (sequence and issue) if the page_count_uncertain flag is set.