http://dev.comics.org/reviews/r/213/
This adds field-by-field comparison pages for publisher and series (copy/paste for the files, but I'll refactor the common parts out later).  This is now what you get when you click on links in the pending/reviewing queues.  There is a link on the comparison pages to pull up the preview in a new window (I normally don't like new window things, but you will almost always want that here, I think- opinions?).

Any logged in person can see the comparison pages, and add a comment.  Only the approver can approved/disapprove/reject, and only when the state is REVIEWING.

The Editing queue now shows all of your changes in the OPEN, PENDING or REVIEWING states, and presents discard, retract, or no action buttons accordingly.