http://dev.comics.org/reviews/r/219/
* Status bar now understands the REVIEWING state and its template code is now properly factored out.
* Modification times on revisions are now stamped correctly (baseline revisions do not need independent modification times, just independent creation times).
* Series preview pages now display the correct first issue cover and link to the correct cover gallery

http://dev.comics.org/reviews/r/220/

This adds creation and workflow stuff for issue and story revisions (i.e. creating the story revisions along with the issue revisions and making sure they track through the process for things like moving from open to pending to reviewing and back- saving updates is left separate, at least for now).

It also ads the edit, compare and preview screens for issues, but *not* for stories- for now they just get listed on the issue page with "edit" buttons that don't go anywhere