Add SeriesRelativeBond for sorting.
This provides a way to sort bonds along a series that may be on either
the origin or target end of any given bond in the list.

A method is provided on the Series model to get a list of SeriesRelativeBonds
wrapping each of the bonds in both directions.  This is not sorted
automatically as some uses may not need the relatively expenseive sorting.

Despite sorting being relatively expensive, with current usage series
never have very many bonds (I know of series with up to five) so doing
the complex sorting in code rather than trying to find a way to keep
the directional bonds *and* sort them together seems like the best option.

SeriesRelativeBond also provides information about whether isues were
specified and provides defaults to allow writing cleaner calling code.