Django 1.5: Remove func generic views from urls.py
All just replaced with the analagous class-based generic views.
In Django 1.8 the patterns() function and the prefix functionality
it provides in the first argument are deprecated, so where I had to
mess with that anyway I went to an empty prefix.  One less thing to do later.

The only non-urls.py function-based generic view import has been removed as
it was no longer being used.