Explicit False default for all BooleanFields.
As of Django 1.6, the database-level default value for BooleanField is None,
which is not a valid value for the field.  As we use a checkbox form widget
for these values, an omitted value should always mean False, which was the
prior default behavior, unless we had already explicitly set default=True.

Also, standardizd on not passing blank=True for BooleanFields, as it seems to
be unnecessary and the docs mostly mention blank=True for text or numeric fields.

Department of lulz:  There was still a really old default=0 from back when
Django's MySQL support didn't quite align MySQL's integer-based Boolean
implementation with the notion of a BooleanField object.  Fixed that.