Updates to get front page to load in Django 1.6
These are the minimal updates necessary to get the site to load under Django 1.6.
Mostly this is due to some minor moving about of imports, plus updated settings
from 3rd-party libraries.

Note that the changes we had made in our fork of django-taggit are now handled
in the more recent version, if the appropriate settings are set (as is done in
this change in settings.py)

There is also a mysterious bug of a search form field accidentally putting a named
argument in as positional, which apparently used to work but now ends up assigning
a string to "max_length".