Use property as a decorator.
The GCD was originally written under Python 2.4, and we just continued
with the old-style property calls for consistency (or due to inertia,
take your pick).

But we're on Python 2.7 these days, so this moves all but two properties
to the decorator syntax introduced in 2.5.  The two kept are a case where
two different properties are implemented using the same private method.