Fix _source and _source_name property helpers.
This was supposed to go with the switch to the property, in which
I dropped the lambda that was making the properties behave in
a polymorphic manner.  This moves to using a helper method for
polymorphism and the main property method as a wrapper in place
of the lambda.

Also, it helps to actually return the result of the helper method
for source_name.