Move CSS-related method to templatetag.
This moves the map for the CSS classes for different changeset states
into the templatetag library for the app that actually needs to use it.

There are still dependencies from gcd on oi, but this change isn't trying
to unravel that.  This is just getting the CSS-related code out of the
model layer, and taking the gcd app CSS dictionary out of the oi app.