Reset migrations and remove creator code (temp)
The creator code is still changing on its branch, so there's no point in
doing anything with it directly in the refactor.  This gets it out of the
way and shortens the amount of time to load a test database.  Both from
dropping the creators tables and from condensing everything down to
one migration.

The creator code will be restored after both it and the refactoring settle.