Re-add stuff to enable writing more tests.
The field lists are needed in order to load the forms module, which I'm
about to do some stuff with.

Also change how GENRES is imported so that the fact that there currently
isn't a GENRES in the OI models doesn't instantly blow up the GCD app.
Non-template code shouldn't be importing templatetags but that's a more
involved part of the refactor.