Custom form field and widget for Keywords.
These are pretty direct knock-offs of the custom form field and widget
that come with the taggit library, but much simpler for our basic
semi-colon split parsing instead of taggit's, which is more flexible but
not compatible with our rules.

Includes unit tests (not everything is mocked, but no database is needed).