Reduce boilerplate for mucking with fields.
We have a great deal of code that just goes down the list of fields
on a type of object and does something with them- copy back and forth
between revisions and display objects, set up the form fields, etc.

This starts streamlining that by consolidating the various lists and
using getattr() and setattr() to just loop over them whenever possible.

As part of this, replaced the old instance method _field_list() with
a class method form_field_list().  As this gets called outside of the
class, it should not have an underscore.  And it does not need an instance.
This will also allow us to drop the get_foo_fields() freestanding functions,
although that's not shown here as there was never a get_publisher_fields().

The form and view modules are not being updated for this at this time.
They're completely broken at this stage of the refactor anyway.