Reorganize fixtures.
This pulls the fixtures out and puts them into a conftest.py file,
which py.test looks for if a test's fixtures can't be found in the test module.

All of the constants become fixtures, and there's more chaining of fixtures
so that fewer things have to be handled separately in the test.

This will make extending the tests to series, issues, etc. easier.