Tests and cleanup for clone_revision()
There was a lot of unnecessary stuff going on with clone_revision.  Possibly
leftover from various things tried during the initial OI development.

This also starts moving towards a more uniform naming for dealing with
fields, i.e. "assignable_fields" instead of "base_fields" which only
really applied to the publisehr-type objects.

In terms of making use of the field lists, as much as possible is being
pushed up into the Revision and RevisionManager classes.