Put model helper functions in separate file.
All of these functions were in models.py before the refactor started, and I've
had to add several back.  To de-clutter the models.py file, I'm restoring all
of them but putting them in a different file.  They will all end up living
in various place, but this makes it easier to work with for now.  There will
not be a "helpers.py" file permanently as the name is too generic, but it
works for the moment.