Add "added" and "edited" properties to Revision.
Various conditionals become a lot easier to read when they are
in the form of self.added or self.edited (symmetical with self.deleted,
which is an actual database column).  As it is, we're inconsistent about
exactly what we check at different times to decide these cases.

Note that TODO about how added doesn't quite work post-commit.  At the
moment that is not a practical problem, but should probably be revisited.