Consolidate various conditionals in a method.
The commit_to_display code has lots of conditions checked about which
things changed and in what direction.  And also very specific ordering
of field assignments in order to have access to the right old and new
values at the right time.

This pulls all of that out and puts it into easy booleans, and saves off
all of the old and new values.  It should be called at the beginning of
commit_to_display and then used as needed through the rest of the method.

This change adds the base method and docstring, and the concrete
method for SeriesRevision.