Revision method for general stat updates.
Most objects have to update general stats, often along with country and/or
language stats.  Provide a method to handle this in the base class.
Subclasses can override and call the superclass before adding more
(typically local count columns like issue_count on the Series table,
not further CountStats).

Subclasses should determine when and whether to call this.