previous_revision and committed for all Revisions
This just promotes the previous_revision OneToOne field that already
existed on a few Revision classes up to the base Revision class, which lets
us drop the old Revision.previous() and Revision.posterior() methods.  Those
methods used the field if it was present, or simulated it if not.

Additionally, this adds a "committed" boolean field to all Revisions.  We have
generally figured out if a Revision is committed based on the changeset state,
but this is insufficient when trying to figure out which revisions within
a changeset have or have not been committed.

Both of these will allow for efficiently implementing Revision.added properly.