Drop unused fields from gcd tables.
Series.publication_notes, Publisher.parent_id, Publisher.is_master,
and Brand.parent_id are all no longer used and do not have a value
for any non-deleted row in production.  The fields are preserved in
the oi revision tables, so historical values, including the final values
in deleted rows, are preserved.

Also, Series.issue_count did not list a default of 0, which all other such
count fields list.  Due to how MySQL handles non-null integer fields, that
was its default in practice anyway, but this makes it clear and removes any
need to initialize it in new Series instances.