Add tests for BrandGroup, Brand, BrandRevision
This does not yet deal with the image resource involved in the Brand, but
otherwise covers the same cases with these revisions that are covered
for PublisherRevision and SeriesRevision.

There is a lot of copy-paste, but my attempts to parametrize the tests for
all of the publisher-like objects did not work out due to how they would
need to be parametrized across the results of calling different fixtures.
Which is not something that works as far as I can tell.

To get these to run, also fix an outdated method call in the Revision classes
and restore the issue_count property on BrandRevision.

Since we ultimately want to move things like the issue_count property out
to the forthcoming Preview objects, this will get changed later, which
should be a pretty natural outcome of the CountStats refactor work.