Fix the RevisionLock unique together constraint.
The docstring comment correctly states that the content type and data object
need to be unique together in order to get a lock on the data object.

But the code used the data object and the changeset for the constraint,
which is wrong as data objects of different types but with the same database
id will fail to lock.  Added a test for this and other conditions.

The changeset is informational here, and may not even be known at the time
that the lock is created (if lock creation fails, we don't create a changeset).