Fix property weirdness, add source_class.
source_class provides the actual data object class that the Revision
works with.  This is needed in some refactoring I'm about to do
where the Revision class will instantiate newly added data objects
without knowing their exact class.

I'm not trying to implement this for ReprintRevision right now as I hope
to greatly simplify that area before moving it to the refactored code.

For reasons I can't recall, we always had this weird double-method call
for the properties on Revision.  I have a vague memory of some scenario
in which overriding a property directly did not work as expected, so
the solution was to delegate the work to a regular method.

I also vaguely recall learning about this on some other project.  But I don't
know if Python behavior changed (possible if unlikely) or if it was
some weird misunderstanding from the days of My First Python App(tm).

Either way, it doesn't make a bit of sense now.  Just use properties normally.