Add and document field declaration methods.
These classmethods on Revision specify how subclasses should declare
their fields according to different desired behavoirs.

This is in preparation for moving clone_revision out of the Manager
class and into a classmethod on Revision.  This will remove the
need for the Manager to know which Model it is managing, which is
in general something that managers should not know as they can
manage many different models.

Currently, Managers do not need to do this, but for the next step
in refactoring Revision creation, they would if we left clone_revision()
where it is.