Refactor RevMgr.clone_revision() to Rev.clone()
This moves clone_revision from the RevisionManager to a classmethod
on the Revision base class.  Among other things, this allows us to
delete all of the child manager classes, substantially reducing
the amount of boilerplate required for adding a new revision type.

Change _copy_assignable_fields_to() to _copy_fields_to() as it does
more than just handle the simple assignable fields.

Make keyword handling part of the field declaration system instead
of it being a weird defaulted parameter to remember.

Fix an error in how new data objects are saved in their first
revision that left the revision unconnected to the data object.