Package updates for Django 1.9
Rather than having some versions specified and some not, this is just
a pip freeze of what got installed when I took the versions off of
everything and did a fresh install into a new virtualenv.

I also had to upgrade setuptools and pip, but have not included them
here as that should be handled in the vagrant environment setup.
Putting the here would still cause some installs to fail the first
time through.