Take code out of gcd/__init__.py
The startup mechanism for Django changed in 1.8 and 1.9.  Now, if your app's
__init__.py has code that directly or indirectly imports model classes, you
will likely not be able to start Django because that code will try to
run before the imported model is configured, causing an exception.

In this case, the 'gcd' app gets configured first, but imported from the 'oi'
app, which was not yet configured.  It's a known bug that gcd relies on oi
at all (that is the wrong direction), but it's best to fix this part of the
problem as well.  The dependency will get fixed later in the refactor.

Anyway, I split the few contents of that file off into their own file.
Each class was only used in one place.