Helper class for related fields and objects.
RelPath (Relation Path) collects code used to work with the sort of relation
field name tuples used in Revision model field behavior descriptions.  Many
of these descriptions return tuples of multiple relations, such as
('brand', 'group') and ('series', 'publisher'), both of which can be applied
to Issue and IssueRevision.

This adds the basics of the class, which only deal with classes and fields.

Subsequent commits will add the ability to get and set values for these
fields when given an instance of the class.  These methods will be a more
generalized form of getattr() and setattr(), and allow us to affect
fields/objects that are more than one Python attribute away from the
actual instance with which we are working.

This will be required for generic code to update statistics, counts, and
other dependent information.