Add Revision._classify_fields() and excluded fields.
Revision._classify_fields() is the core of the generic revision / data object
field management code.  It uses the Model._meta API introduced in Django 1.8
to inspect the fields on the data object and the revision and automatically
put them into various categories previously set up by defining explicit field
sets in each child class.

The _get_excluded_class_names() method allows for filtering certain data object
fields out of the classifications entirely when they cannot be handled
generically.  This method is the first of several to use the new naming scheme
where all of the methods for accessing information on fields start with _get,
and the final portion describes the format:

_get_*_fields() for dictionaries of names to field objects
_get_*_names()  for sets of names (strings)
_get_*_tuples() for tuples of successive names, i.e. ('series', 'publisher')
                to access Issuse.seris.publisher

Other classification methods will be moved to the new naming scheme as they
are updated to work with _classify_fields().