Standardize the classified field methods.
This adds the methods that trigger field classification and access the
results, as well as updates the remaining manual declaration methods
to a standard naming scheme that indicates the format of their return
values.

With auto-classification, there is no longer need to manually declare
assignable, non-assignable, and many-to-many fields.

With the RelPath helper, it should also no longer be necessary to
distinguish between foreign key parents and many-to-many parents.

Originally, I was checking for classification in all of the access
methods, but that was noisy so I just moved it into _classify_fields().
This is not a high-performace part of the code so an extra method call
is fine.

Finally, I moved all of the field related methods up to the top of the
unit tests along with the init and classification tests.  The class
methods also can be run directly on Revision rather than DummyRevision.