Add data object base classes.
GcdData and GcdLink collect common fields and define methods that
are expected to appear on every child object.  Note that the modified
property and deleted data member for GcdLink came from SeriesBond and
were not present in the reprint classes.  I think having them here
is harmless, especially as we will be consolidating the reprint classes
into one class before the refactor is complete.

The schema migration just standardizes on the nullability and data indexes.
Image.created and .modified become NOT NULL, but in practice they are never
null anyway so this won't cause problems (because auto_now and auto_add_now).

There are also some flake8 compliance changes here, notably in seriesbond.py

Finally, in Django 1.8 or 1.9, django.form.util became django.form.utils,
and I needed to mark a bit of mess with TODOS in apps.oi.views (which is
currently completely broken anyway, so I didn't bother trying to fix these yet).