Add database system tests for SeriesBonds.
This is mostly the same set of test cases that we've been using for the other
revision classes, but with a test for deletion added since SeriesBonds
handle deletions differently, and refactoring them is mostly about moving
that code to the generic code.

Bonds attached to issues aren't covered because there's a lot of work
to be done to set up issues for database tests.  That will be added later.