Non-db unit tests for IssueRevision.
Also, after getting confused for the millionth time, moved some stray
database fields in Issue that were listed after some properties back up to
be declared alongside all of the other fields.

Restored a missing utility method needed to get the tests running,
and stubbed out a soon-to-be-obsolete Changeset method.

Put the INDEXED constants into the main apps.gcd.models namespace, so that
it fits better with our relatively recent approach to how those should
be imported in various parts of the code (i.e. through apps.gcd.models
unless you are arlready in a sub-module of apps.gcd.models).