Add tests for more issues and non-comics issues.
This adds tests for adding additional issues to a series either with
the default positioning (at the beginning) or with an explicit position.

It also adds a long test going through the lifecyle of a non-comics issue
with variant to ensure that all of the accounting is handled properly
in these cases.  Other aspects of the lifecycle are assumed to work in
the way shown by the other cases, as they are not controlled by the
comics-ness of the series.