Remove superflous init file- it breaks tests.
Our top level is not really a python package.  Nor (based on current packaging
recommendations on docs.djangoproject.org) should it be.  Having this
__init__.py file here confuses py.test into importing things both starting
with "www.apps" and "apps" as the packages, which means it finds things
(particularly conftest modules) twice.  This is bad, as it gets confused
about redefinitions.

Since it's causing problems and adds no benefit, delete it.