Handle multi-valued parent updating.
Expose multi-value-ness in RelPath.

Move dummy revision classes into conftest because you can only patch
things if you import them, and the behavior of model class fields
means that (as far as I can tell) patching is the only way to get
certain things to work without database access.

Add unit tests for old and new parent count adjustment functionality
(I think the old functionality did not have a test in test_revision.py
because it was covered by test_series_revision.py, but it's best to
have specific tests in the right module for it- I think I meant to
come back to this anyway).