Fix error in walking fields / classes in RelPath.
A key spot forgot to use ".rel" to look at the field and model on the far
end of the relation so that we properly walk the relations / models out.
This fixes that and ensures that the test will catch it.

Also, use _first_model_class instead of _model_classes[0].  This was the
original design but got forgotten somewhere along the way.  Using
_first_model_class works with the non-relational support also added in
this change, while _model_classes[0] does not because it is only
populated when relational fields are found.