Special stat for series issue counts. Fix deletes.
The local cache count of issues that Series maintains tracks non-variant
issues whether or not the series is a comics publication.  This makes sense
but is different from all other issue count behavior.  Add a special stats
value to handle it, which regrettably requires a special case in CountStats
to ignore the special case for Series.

The alternative would be a more complex stat counts dictionary, which seems
overkill for now.

Also ensure that deleted objects always return empty stat counts, and add
tests for this.  BrandUse does not get a test because it only has a local
stat_counts() method because its base class still needs to be sorted out.