Intelligent country/language for stats.
So that whole "there is no need for subclasses to override this" comment
for getting the field for figuring out the country and/or language relevant
to a stat was totally wrong, but at least I had the sense to make it a method
anyway.

This moves the logic for handling "missing" (i.e. irrelevant) local country
or language fields into the method that figures out the fields, which makes
more sense than handling it at the point of use.  It also makes this method
return tuples, so we don't have to convert the output to tuples anymore.

Finally, it has to be tuples because lower objects figure out their language
and country by looking further up the hierarcy (specifically to series),
so add those overrides and tests for all of it.