Field copy shouldn't save, edited should be bool.
The field copying method saves in order to do m2m/keyword things.  I vaguely
remember doing that and not liking it, but I guess forgot?  The m2m is already
handled in the correct place in commit_to_display(), so move the keyword
stuff there and remove target.save() and the duplicate m2m code.

Also, ensure the edited property is always a bool (like added and deleted).