Add helpers for IssueRevision refactor.
It simplifies the upcoming refactored code to pull out some database queries
into helper functions or properties.  Determining whether the series changed
(indicating an issue move) seems useful in general so it is a public property
available to the UI and other code.

The various private methods help us write code that handles bulk adds, deletes
and moves without having to rely on CTYPES to tell us what is going on.  This
will be made more clear in upcoming usage, but the basic idea is that if we
are one of several issue add revisions in a changeset on the same series, then
we must be in a bulk add situation.  We need to find the revision that specifies
"after" (which should be the one with the earliest revision sort code, but we
check for that in other code that isn't in this change).

We also need to sort the revisions and process them correctly.  If we are trying
to add (or move, which looks like an add to the target series) a revision that
is not the first issue in the run, we will need to back up and first commit the
earlier issues in the run (_open_prereq_revisions).  Later, we will need to
look at the most recently committed revision (using _committed_prereq_revisions)
in order to figure out the issue after which we want ot add the current revision.

These work with deletes as well by reversing the order (deleting from last to
first).  Since we currently don't actually delete issues from the database, this
is not strictly required, but is nicely symmetrical and feels more correct than
deleting from first to last.