Finish migrating IssueRevision.commit_to_display()
Handle special case fields, including calculating the sort_code for added
or moved issues.

Handle stuff that needs to happen when the series changes, such as the series
gaining or losing a cover gallery.

Fold the old _post_commit_to_display() code into the new
_post_adjust_stats() method that is called form within commit_to_display(),
and drop _post_commit_to_display() entirely as this was its only use.

Remove the old first/last issue setting method- the code was already migrated
to Series, but had not been deleted as it was still being called here.