Standard GcdLink fields, better state handling.
The GCD app should ideally not be aware of the OI app.  That's a bit much
to ask right now, but we should at least stop importing modules from the OI
which causes circular dependency problems.

This moves they typical logic that requires knowlege of state values
into the RevisionManager class methods (and the new RevisionQuerySet class)
so that GCD code now only needs to know those methods, and not the exact
nature of the states.

Additionally, standardize the created/modified fields for GcdLink objects,
and drop the obsolete open_reserve field on Series.

Finally, move the basic deletable() method into the base class, and switch
the additional code over to a has_dependents() method that only a few
classes need to override.

Use exists() instead of count() == 0 as it is faster and more readable,
and results in the correct type (bool).

Have the tests check boolean results with "is" to ensure that they actually
return bools.