Expand Reprint model to handle all reprints.
We want to merge the IssueReprint, ReprintFromIssue, and ReprintToIssue
models into the Reprint model, which will then work like SeriesBond.

Add the new fields needed, but leave them nullable and with a non-clashing
reverse relation name for this step.  Also add code to ensure that the
issue fields are always set consistently.

Fix an apparent bug where the compare string in some cases swapped the
issue URL and full name.  Also standardize the unicode form a bit,
and (hopefully) improve the readability of the issue form.