Update for one Reprint table & new Revision flow.
This drops the fields related to managing the four different reprint data
tables, renames two of the remaining fields to match the equivalent Reprint
fields and work with the generic Revision code, and drops all of the old
code in the class.

In its place, we have one commit customization method, which handles the
case of linking to story revisions instead of stories, and needing to
ensure that those revisions have already been committed.

In addition, save() is customized in a manner somewhat similar to Reprint.save(),
but with the assumption that this code cannot know what an indexer meant, and
must raise an exception if anything conflicts.  This would result in the
edit form showing an error in the UI, which the indexer could then fix.