Remove the three obsolete reprint tables.
These have been consolidated into the Reprint table.  This does not deal
with code that is broken by this change beyond ensuring that we can pass
a "python manage.py check" and pass unit tests.

The new reverse relations are different from any of the pre-existing ones,
which are now simulated by properties.  This, of course, will not work for
database filters, but that will be addressed in future commits.