Better names for related revision hooks.
_pre_stats_measurement() and _post_adjust_stats() were, like the other
commit hooks, named for their position in the process.  However, they really
have a clear function for handling related revisions before and after the
current commit, so rename them _handle_prerequisites() and _handle_dependents().

Note that prereqs or dependents may be things other than related revisions, but
that is the current and generally expected usage.

Also clarified the difference between _handle_prerequisites() and
_pre_commit_check().

In the process of updating the documentation, realized that we never
set the committed flag properly when we commit a revision, so fixed that
and the related tests.

This further uncovered that for LinkRevision subclasses, the unhooking
from the to-be-deleted GcdLink object was not affecting the current
in-memory instance, so when it was later saved it tried to re-connect
to the now-deleted GcdLink.  Fixed that as well.