Migrate StoryRevision to new system.
Change Issue.set_index_status()'s return value to be a delta for use in
count stats adjustments.  The old return value was not being used anywhere else.

Remove an old comment in IssueRevision about which things are in which method,
because we've since renamed the methods such that the comment no longer makes
sense anyway.

I changed the imports in countstats because I was trying to resolve a circular
import.  I came up with a better solution that did not have that problem, but
left the imports changed to relative as they now match the other within-package
imports for gcd.apps.models.*