"meta" field classification, minor housekeeping
Introduce "meta" fields for things like marked (and eventually the
corner coordinates for covers and probably other stuff).  See docstrings
for details.  These are not copied from data to revision, but are copied
from revision to data.

Removed "changes" from _copy_fields_to() because the only thing that referenced
it was a TODO comment and I couldn't think of any use cases for continuing
to pass it.

Minor tweaking of some comments to clarify the organization of methods in
the Revision base class.  Removed stray commented-out debug code.