Add IssueRevision.fork_variant().
This moves the logic for how to build a typical variant from an existing
issue inside of the IssueRevision class.  While most revision construction
should stay in the views, in this case the construction has a lot more to
do with the internal field (and StoryRevision) behavior of variants than
it does with the input from the view.  Only the variant name and the
request for a reservation are passed down directly from the form.