Series links on issue page, nicer disabled deletes
Rename Resources to Related Scans and make the nature of the scans
more clear.

Move the change history link up to the bottom of the edit box, and
styles it as a footer to that box.  The "View Change History" link
(like the my.comics.org link) is styled centered.

The large side column headings are now all non-links, and styled to align left.

This also replaces the wordy and disconcertingly yellow "can't delete" box
with a simple disabled button, with minimal explanatory text that links to
a page on the documentation site explaining why it is disabled in detail.

The yellow made it look like something was wrong, which is incorrect- being
unable to delete objects is a normal condition.  Originally, the yellow
can't-delete panel couldn't appear on issue pages, and only appeared on
the rather less-frequently viewed other page types.  However, with the advent
of variants, it is now common that an issue cannot be deleted, increasing
the usage of the relatively noisy yellow box.  This version reduces the
noise in that corner of the page.

Various CSS tidying and consistency stuff.

Also, fix ads. This matches the ad location with the current locations
on production.