Use new templates for messages.
This includes CSS styles for the special and survey banners, such that
when those banner templates are empty (commented out) they do not affect
the page layout.

The indexer notice CSS change ensures that there is space before the bottom
elements of the page even when the queue is empty.