Package updates for Django 1.6 support.
Since this part of the branch of forward migraiton was last worked on, the
django18 branches of our customized repos moved to later versions of Django.
As one might guess from the "django18" name.  Now that we have decided to
do some incremental deploying, point this file at a specific "django16"
branch for each package.

Also, update django-compressor to the minimum version necessary for
supporting Django 1.6