Python side of conversion to staticfiles.
Use 'static/' as the general dir to collect *from*, and 'static_root/'
as what to collect *to*.  There will no longer be a 'media/' directory,
just the 'site_media/' which is entirely populated by contributed
content and therefore ouside of version control.

Note that both MEDIA_URL and STATIC_URL are sent to templates via
context managers, so I removed any explicit passing of these.
Passing 'media_url' predates the existence of context managers and
is long since obsolete.  Probably dates to Django 0.96.

Finally, there is a special view that now does the debug-only static
file serving that we had set up in a slightly more complex way before.