Move help text out to forms, separate out data.
Split big data structures (of strings, mostly) out into a support module
to reduce clutter.  This might not be the final organizatino of this
sub-package but it helps for now.

Help text is a UI concern and should not be entangled with model code.
Django has introduced the help_texts dict for forms in the years since
this code was originally written, so we'll take advantage of that now.

Likewise, verbose_name is only being used for form labels, so just
do that directly.  We don't need friendly admin UI labels which is
the other place verbose_name seems to get used.  We do need to
title case the first word of the labels (which used to be done
automatically when verbose_name values were copied over by Django).

In order to get the help texts and labels in all of the right places, the
easiest thing to do is make the Meta classes inherit in parallel with the
enclosing form classes.  And make sure that the base of each hierarchy
with labels/help_texts has them assigned directly.