Replace raise False with return False.
With Django 1.7, attempting to raise a boolean produces a TypeError.
It's not clear why it did not before, as that should be a Python thing,
but the code definitely goes down that path under 1.6 as well and is fine.

It is not entirely clear if there are situations where we relied on the
exception throwing behavior.  The call that uncovered the bug was a
direct call from the templates.