Simplify query, use wide widget for more fields.
For some reason, the old way this query was written stopped working on
Django 1.7 due to failing a renaming/aliasing check deep in the database code
when it appeared to rename things to themselves, in otherwords not rename
at all.  This seems like it might have been a Django 1.7 bug but this
version of the query using Q objects works and is easier to read anyway.

Also, make some text fields wider as requested on gcd-main.