Make help widget visible, put help in right place.
Prior to Django 1.7, a widget's is_hidden field was a simple boolean that
was set independent of anything else.  In 1.7, it is now a read-only
property that is completely tied to the HTML input type.  So our
HiddenInputWithHelp stopped working.  The solution is to override
the property to return False for HiddenInputWithHelp.

Also, the help text of the on_sale_date_help field was being set
at the wrong point in the form initialization by having it off with
the other help texts.  This moves it into the field declaration,
which might not be ideal but we are not (yet) consistent about how
we set these things anyway.