Don't make gcd-managed-content a submodule.
The managed content is expected to change in production.  If it is a
submodule, then it will constantly show itself as being out of sync
in the parent repository.  Since we can now conditionally include
templates and default to not using the managed content, let's just
make this a gitignored directory instead, and most development
environments can function without it entirely.