Move Error model into indexer app.
While not an entirely satisfactory location, this is better than any of the
currently available alternatives, and making an app just for errors did not
seem quite right.  Errors mostly happen because of user/indexer actions,
so it has at least a bit more affinity for the indexer app than the
purely data-oriented gcd app.