Arrange fixtures as one per table.
This allows us to selectively populate or reset specific tables rather than
having to do several or mess with the files.  It is also important to be
able to conditionally check for contents when running these migrations
immediately after updating to Django 1.7.

Also, move fixtures as needed to reflect the new app organization.