Update tools and retire obsolete history scripts.
Add the Django 1.7+ initialization call to the scripts we
still need to use.  Delete scripts from "migration" which fixed one-time
things in the old database, particularly for object types which are fully
migrated (series and publishers).

Also the autofill keydate script (which might still be useful?)
and calculate_imps (which is presumably needed for the last bit of
migration?)