Add py.test, remove unneeded top level init file.
This includes the py.test packages and the packge to make it play nice
with Django, and .gitignores for py.test's config file and cache directory.

The top level __init__.py confuses py.test by making it think that module
paths such as www.apps.gcd.models are how are tree is structured.
Which is incorrect- our top level directory is not a python package or module.
So just delete it.