Move award initial data to new migration.
This is needed to avoid the following problem when starting from an
empty database:

Running migrations:
  Applying gcd.0004_initial_creator_data...
  [...]
  django.db.utils.OperationalError: Problem installing fixture
    'apps/gcd/fixtures/award.json': Could not load gcd.Award(pk=1):
    (1054, "Unknown column 'created' in 'field list'")

The modified award.json contains all current awards from production.

On an instance where the award data already exist, we might need to do:

./manage.py migrate --fake gcd 0007_initial_award_data

to mark the migration as already applied (although it shouldn't matter).